requirements.test.txt
hypothesis==1.14.0
pytest-cov==2.2.0
coveralls==1.1
